+++ inst/include/mvnorm.h
@@ -4,7 +4,15 @@
 // [[Rcpp::depends("RcppArmadillo")]]
 #include <RcppArmadillo.h>
 
-double dmvnorm(arma::vec const& x, arma::vec const& mean, arma::mat const& Sigma, bool log = false);
+double dmvnorm(
+    arma::vec const& x, arma::vec const& mean, arma::mat const& Sigma,
+    bool log = false
+);
+
+double pmvnorm(
+    arma::vec const& x, arma::vec const& mean, arma::mat const& Sigma,
+    double abseps = 1e-3
+);
 
 arma::vec rmvnorm(arma::vec mean, arma::mat const& Sigma, bool log = false);