+++ src/mvnorm.cpp
@@ -1,5 +1,6 @@
-// [[Rcpp::depends("RcppArmadillo")]]
+// [[Rcpp::depends(RcppArmadillo, mvtnorm)]]
 #include <RcppArmadillo.h>
+#include <mvtnormAPI.h>
 
 // [[Rcpp::interfaces(cpp)]]
 // [[Rcpp::export]]
@@ -16,6 +17,51 @@ double dmvnorm(
   double log_det = 2.0 * arma::sum(arma::log(L.diag()));
   double log_density = -0.5 * (p * std::log(2.0 * M_PI) + log_det + quad);
   return log ? log_density : std::exp(log_density);
+}
+
+// [[Rcpp::interfaces(cpp)]]
+// [[Rcpp::export]]
+
+double pmvnorm(
+  arma::vec const& x, arma::vec const& mean, arma::mat const& Sigma,
+  double abseps = 1e-3
+) {
+  int n = x.n_elem;
+  arma::vec sds = arma::sqrt(Sigma.diag());
+  arma::mat corrmat = arma::diagmat(1.0 / sds) * Sigma * arma::diagmat(1.0 / sds);
+  arma::vec bound = (x - mean) / sds;
+  arma::vec lowertrivec(n * (n - 1) / 2);
+  int k = 0;
+  for (int i = 0; i < n; ++i) {
+    for (int j = 0; j < i; ++j) {
+      lowertrivec(k++) = corrmat(i, j);
+    }
+  }
+  int nu = 0;
+  int maxpts = 25000;
+  double releps = 0;
+  int rnd = 1;
+  double* bound_ = bound.memptr();
+  double* correlationMatrix = lowertrivec.memptr();
+  double* lower = new double[n];
+  int* infin = new int[n];
+  double* delta = new double[n];
+  for (int i = 0; i < n; ++i) {
+    infin[i] = 0;
+    lower[i] = 0.0;
+    delta[i] = 0.0;
+  }
+  double error;
+  double value;
+  int inform;
+  mvtnorm_C_mvtdst(
+    &n, &nu, lower, bound_, infin, correlationMatrix, delta, &maxpts, &abseps,
+    &releps, &error, &value, &inform, &rnd
+  );
+  delete[] (lower);
+  delete[] (infin);
+  delete[] (delta);
+  return value;
 }
 
 // [[Rcpp::interfaces(cpp)]]